+++ lib/punchblock/translator/asterisk/component/mrcp_recog_prompt.rb
@@ -45,6 +45,7 @@ module Punchblock
             raise OptionError, "An initial-timeout value must be -1 or a positive integer." if @initial_timeout < -1
             raise OptionError, "An inter-digit-timeout value must be -1 or a positive integer." if @inter_digit_timeout < -1
             raise OptionError, "A recognition-timeout value must be -1, 0, or a positive integer." if @recognition_timeout < -1
+            raise OptionError, "A speech-complete-timeout value must be a positive integer."  if @speech_complete_timeout < -1
           end
 
           def execute_app(app, *args)
@@ -60,6 +61,7 @@ module Punchblock
               opts[:ct] = input_node.min_confidence if input_node.min_confidence
               opts[:sl] = input_node.sensitivity if input_node.sensitivity
               opts[:t]  = input_node.recognition_timeout if @recognition_timeout > -1
+              opts[:sct] = input_node.speech_complete_timeout if @speech_complete_timeout > -1
               yield opts
             end
           end
@@ -68,6 +70,7 @@ module Punchblock
             @initial_timeout = input_node.initial_timeout || -1
             @inter_digit_timeout = input_node.inter_digit_timeout || -1
             @recognition_timeout = input_node.recognition_timeout || -1
+            @speech_complete_timeout = input_node.speech_complete_timeout || -1
           end
 
           def grammars