+++ spec/punchblock/translator/asterisk/component/mrcp_prompt_spec.rb
@@ -612,6 +612,36 @@ module Punchblock
             end
           end
 
+          describe 'Input#speech-complete-timeout' do
+            context 'a positive number' do
+              let(:input_command_opts) { { speech_complete_timeout: 1000 } }
+
+              it 'should pass the dit option to SynthAndRecog' do
+                expect_synthandrecog_with_options(/sct=1000/)
+                subject.execute
+              end
+            end
+
+            context 'unset' do
+              let(:input_command_opts) { { speech_complete_timeout: nil } }
+
+              it 'should not pass any options to SynthAndRecog' do
+                expect_synthandrecog_with_options(//)
+                subject.execute
+              end
+            end
+
+            context 'a negative number other than -1' do
+              let(:input_command_opts) { { speech_complete_timeout: -1000 } }
+
+              it "should return an error and not execute any actions" do
+                subject.execute
+                error = ProtocolError.new.setup 'option error', 'A speech-complete-timeout value must be -1 or a positive integer.'
+                expect(original_command.response(0.1)).to eq(error)
+              end
+            end
+          end
+
           describe 'Input#mode' do
             pending
           end